Cleanup some toast and nav stuff
* Removed unnecessary icon customization, just use built in close for
  toast notifications.
* Toasts were hanging around too long, progress bar provides better
  notification experience.
* No need for a toast styled component.
* Move StagingBanner to utilities styled components.
* Make BannerWrapper not need 2 components.
