--- src/common/components/Notification/ErrorNotificationIcon.tsx
@@ -1,5 +1,0 @@
-import { faTimesCircle } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { FC } from 'react';
-
-export const ErrorNotificationIcon: FC = () => <FontAwesomeIcon icon={faTimesCircle} />;
