--- src/common/components/Notification/SuccessNotificationIcon.tsx
@@ -1,5 +1,0 @@
-import { faCheck } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { FC } from 'react';
-
-export const SuccessNotificationIcon: FC = () => <FontAwesomeIcon icon={faCheck} />;
