--- src/common/components/Notification/index.ts
@@ -1,3 +1,0 @@
-export * from './NotificationContainer';
-export * from './ErrorNotificationIcon';
-export * from './SuccessNotificationIcon';
