+++ src/common/services/notification.tsx
@@ -1,15 +1,13 @@
-import { ErrorNotificationIcon, SuccessNotificationIcon } from 'common/components/Notification';
 import { toast } from 'react-toastify';
 
 export const showSuccessMessage = (message: string): void => {
   toast.success(message, {
-    icon: <SuccessNotificationIcon />,
+    toastId: message,
   });
 };
 
 export const showErrorMessage = (message: string): void => {
   toast.error(message, {
-    icon: <ErrorNotificationIcon />,
     toastId: message,
   });
 };
