+++ src/features/navbar/components/VerticalNav.tsx
@@ -1,55 +1,15 @@
+import { CustomSelect } from 'common/components/CustomSelect';
+import { BitwiseNavbar } from 'common/styles/page';
 import { useAuth } from 'features/auth/hooks';
 import { FC } from 'react';
 import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
-import styled from 'styled-components';
+import { useTranslation } from 'react-i18next';
+import { languages } from '../../../i18n/config';
 import { useLogoutModal } from '../hooks/useLogoutModal';
 import { useNavLinks } from '../hooks/useNavLinks';
 import { CustomNavAction, CustomNavLink } from './CustomNavLink';
 import { Logo } from './Logo';
 import { NavUserDetails } from './NavUserDetails';
-import { CustomSelect } from 'common/components/CustomSelect';
-import { useTranslation } from 'react-i18next';
-// import Button from 'react-bootstrap/Button';
-import { languages } from '../../../i18n/config';
-
-export const BitwiseNavbar = styled(Navbar)`
-  background: ${props => props.theme.nav.backgroundColor};
-  align-items: flex-start;
-  padding: 2rem;
-  overflow-y: auto;
-  width: 280px;
-  height: 100vh;
-  z-index: 1;
-  position: fixed;
-  box-shadow: 1px 0 0 0 #dadada;
-
-  .navbar-brand {
-    padding-top: 0;
-  }
-
-  .navbar-brand > img {
-    width: 64px;
-    margin-left: 1rem;
-    margin-bottom: 2rem;
-    opacity: 0.9;
-    margin-top: 1.5rem;
-  }
-
-  .nav-wrap {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-
-    .navbar-nav:first-of-type {
-      flex: 1;
-    }
-
-    .navbar-nav:nth-of-type(2) {
-      margin-bottom: 1rem;
-    }
-  }
-`;
 
 type Props = {
   closeVerticalNav?: () => void;
